Moves to internals.interfaces
