+++ neuro_san/internals/graph/tools/external_tool.py
@@ -17,8 +17,8 @@ from typing import List
 import json
 
 from neuro_san.interfaces.async_agent_session import AsyncAgentSession
-from neuro_san.interfaces.async_agent_session_factory import AsyncAgentSessionFactory
-from neuro_san.interfaces.invocation_context import InvocationContext
+from neuro_san.internals.interfaces.async_agent_session_factory import AsyncAgentSessionFactory
+from neuro_san.internals.interfaces.invocation_context import InvocationContext
 from neuro_san.internals.journals.journal import Journal
 from neuro_san.internals.messages.chat_message_type import ChatMessageType
 from neuro_san.internals.run_context.interfaces.callable_tool import CallableTool
