+++ neuro_san/internals/run_context/langchain/langchain_run_context.py
@@ -33,9 +33,9 @@ from langchain_core.messages.system import SystemMessage
 from langchain_core.prompts import ChatPromptTemplate
 from langchain_core.tools import BaseTool
 
-from neuro_san.interfaces.async_agent_session_factory import AsyncAgentSessionFactory
-from neuro_san.interfaces.invocation_context import InvocationContext
 from neuro_san.internals.errors.error_detector import ErrorDetector
+from neuro_san.internals.interfaces.async_agent_session_factory import AsyncAgentSessionFactory
+from neuro_san.internals.interfaces.invocation_context import InvocationContext
 from neuro_san.internals.journals.journal import Journal
 from neuro_san.internals.run_context.interfaces.agent_tool_factory import AgentToolFactory
 from neuro_san.internals.run_context.interfaces.run import Run
