+++ neuro_san/internals/run_context/utils/external_tool_adapter.py
@@ -16,7 +16,7 @@ from grpc import StatusCode
 from grpc.aio import AioRpcError
 
 from neuro_san.interfaces.agent_session import AgentSession
-from neuro_san.interfaces.async_agent_session_factory import AsyncAgentSessionFactory
+from neuro_san.internals.interfaces.async_agent_session_factory import AsyncAgentSessionFactory
 
 
 # pylint: disable=too-few-public-methods
