+++ neuro_san/session/session_invocation_context.py
@@ -12,8 +12,8 @@
 
 from leaf_common.asyncio.asyncio_executor import AsyncioExecutor
 
-from neuro_san.interfaces.async_agent_session_factory import AsyncAgentSessionFactory
-from neuro_san.interfaces.invocation_context import InvocationContext
+from neuro_san.internals.interfaces.async_agent_session_factory import AsyncAgentSessionFactory
+from neuro_san.internals.interfaces.invocation_context import InvocationContext
 
 
 class SessionInvocationContext(InvocationContext):
